# db-assistant-ui/src/app/services/models/db-connection-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DbConnectionRequest {
  database?: string;
  dbType?: string;
  host?: string;
  name?: string;
  password?: string;
  port?: number;
  username?: string;
}