# db-assistant-ui/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenApi specification - marouane",
    "description": "OpenApi documentation for the db-assistant project with spring security",
    "contact": {
      "name": "marouane",
      "email": "marouaneyazza@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local env"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/db": {
      "get": {
        "tags": [
          "Database Connections"
        ],
        "summary": "List saved database connections",
        "description": "Retrieves all database connections saved by the authenticated user. Does not include passwords, but contains metadata such as name, host, port, and type.",
        "operationId": "listConnections",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DbConnectionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Database Connections"
        ],
        "summary": "Create and activate a database connection",
        "description": "Stores a new database connection configuration for the authenticated user and immediately activates it.",
        "operationId": "createAndActivateConnection",
        "requestBody": {
          "description": "Database connection details for creating and activating a connection",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbConnectionRequest"
              },
              "example": {
                "name": "MyPostgresDB",
                "dbType": "postgresql",
                "host": "localhost",
                "port": 5432,
                "database": "db-assistant",
                "username": "user",
                "password": "password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connected and saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbConnectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid connection details or connection failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/db/{id}/activate": {
      "post": {
        "tags": [
          "Database Connections"
        ],
        "summary": "Activate a saved database connection",
        "description": "Activates a database connection previously saved by the authenticated user. If another connection is already active, it will be replaced.",
        "operationId": "activateConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/db/test": {
      "post": {
        "tags": [
          "Database Connections"
        ],
        "summary": "Test a database connection",
        "description": "Tests whether the provided database connection details are correct and the database is reachable.",
        "operationId": "testConnection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "description": "Accepts registration details and creates a new account.",
        "operationId": "register",
        "requestBody": {
          "description": "Registration details including firstName, lastName, email, and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              },
              "example": {
                "firstName": "John",
                "lastName": "Doe",
                "email": "john.doe@example.com",
                "password": "strongPassword123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate a user",
        "description": "Validates credentials and returns a JWT token on success.",
        "operationId": "authenticate",
        "requestBody": {
          "description": "User credentials including email and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              },
              "example": {
                "email": "john.doe@example.com",
                "password": "strongPassword123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/db/{id}": {
      "delete": {
        "tags": [
          "Database Connections"
        ],
        "summary": "Delete a saved database connection",
        "description": "Deletes a database connection previously saved by the authenticated user. If the connection is currently active, it will be deactivated.",
        "operationId": "deleteConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DbConnectionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dbType": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "database": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "DbConnectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "dbType": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "database": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "TestConnectionResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RegistrationRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      },
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          }
        },
        "required": [
          "email"
        ]
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}