# db-assistant/src/main/java/com/marouane/db_assistant/database/TestConnectionResponse.java
package com.marouane.db_assistant.database;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class TestConnectionResponse {
    private String status;
    private String message;
}